Additional hub sensors (#22)
This PR adds a couple of additional sensor for the Sutro Hub.

It adds the hub battery charger status and the SSID the hub is connected to. It also adds the latest message send time to the sensors as an attribute. This is different from the reading taken time which is also appended as an attribute to cartridge based readings.

It also simplifies the device info and adds the current firmware version to the device

Signed-off-by: Brian Towles brian@towles.com